backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
supabase==2.3.6
python-dotenv==1.0.1
pydantic==2.6.4